fix: board tokens

See merge request s08-bigdata-recom-sub2/S08P22A704!164

+++ back-spring/src/main/java/com/ssafy/pnut/controller/BoardController.java
@@ -2,9 +2,7 @@ package com.ssafy.pnut.controller;
 
 import com.ssafy.pnut.common.response.BaseResponseBody;
 import com.ssafy.pnut.dto.*;
-import com.ssafy.pnut.entity.board;
-import com.ssafy.pnut.entity.boardSteps;
-import com.ssafy.pnut.entity.likeTable;
+import com.ssafy.pnut.entity.*;
 import com.ssafy.pnut.service.*;
 import io.swagger.annotations.*;
 import lombok.RequiredArgsConstructor;
@@ -15,6 +13,7 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 import springfox.documentation.annotations.ApiIgnore;
 
+import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import java.io.IOException;
 import java.time.LocalDateTime;
@@ -49,8 +48,12 @@ public class BoardController {
             @ApiResponse(code = 404, message = "사용자 없음"),
             @ApiResponse(code = 500, message = "서버 오류")
     })
-    public ResponseEntity<? extends Object> createRecipe(@RequestPart @ApiParam(value="게시글 정보", required = true) RecipeCreateReq recipeCreateReq, @RequestPart @ApiParam(value="file", required = false) List<MultipartFile> file) {
+    public ResponseEntity<? extends Object> createRecipe(@RequestPart @ApiParam(value="게시글 정보", required = true) RecipeCreateReq recipeCreateReq, @RequestPart @ApiParam(value="file", required = false) List<MultipartFile> file, HttpServletRequest request) {
         try {
+            UserDto userDto = userService.getUserByToken(request.getHeader("Authorization").substring(7));
+
+            recipeCreateReq.setUserEmail(userDto.getEmail());  //  글 작성 유저 정보 저장
+
             String thumbnailImg = awsS3Service.uploadImage(file.get(0));  // 썸네일 이미지 S3에 올림
             file.remove(0);  // 리스트에서 썸네일 삭제
 
@@ -58,6 +61,57 @@ public class BoardController {
             List<String> ImgNameList = awsS3Service.uploadImages(file);  // 레시피 단계 사진들 S3에 게시
 
             boardStepsService.save(recipeCreateReq.getRecipe_steps(), Board, ImgNameList, recipeCreateReq.getStepNums());  // 레시피 단계 글이랑 사진 DB에 저장
+
+            return ResponseEntity.status(200).body(BaseResponseBody.of(200, "success"));
+        } catch (Exception e) {
+            e.printStackTrace();
+            return ResponseEntity.status(200).body(BaseResponseBody.of(401, "Bad Request"));
+        }
+    }
+
+    @PatchMapping("/{boardId}")
+    @ApiOperation(value = "게시판 글 수정", notes = "<strong>게시판 글 수정</strong>")
+    @ApiResponses({
+            @ApiResponse(code = 200, message = "성공"),
+            @ApiResponse(code = 204, message = "No Content"),
+            @ApiResponse(code = 401, message = "인증 실패"),
+            @ApiResponse(code = 404, message = "사용자 없음"),
+            @ApiResponse(code = 500, message = "서버 오류")
+    })
+    public ResponseEntity<? extends Object> patchRecipe(@PathVariable("boardId") Long id, @RequestPart @ApiParam(value="게시글 정보", required = true) RecipeCreateReq recipeCreateReq, @RequestPart @ApiParam(value="file", required = false) List<MultipartFile> file, HttpServletRequest request) {
+        try {
+            UserDto userDto = userService.getUserByToken(request.getHeader("Authorization").substring(7));
+
+            Optional<board> Board = boardService.findById(id);  // 작성한 게시판 불러옴
+            if(!Board.isPresent())  // id에 맞는 게시글이 없으면 null리턴
+                return ResponseEntity.status(200).body(BaseResponseBody.of(401, "There's no such BoardId"));
+
+            if(Board.get().getUserEmail().getEmail() != userDto.getEmail()) {  // 본인이 작성한 레시피가 아니라면
+                return ResponseEntity.status(200).body(BaseResponseBody.of(401, "Unauthorized"));
+            }
+
+            String thumbnailImg = awsS3Service.uploadImage(file.get(0));  // 썸네일 이미지 S3에 올림
+            file.remove(0);  // 리스트에서 썸네일 삭제
+
+            Board.get().setContent(recipeCreateReq.getContent());
+            Board.get().setTime(recipeCreateReq.getTime());
+            Board.get().setTitle(recipeCreateReq.getTitle());
+            Board.get().setQuantity(recipeCreateReq.getQuantity());
+            Board.get().setIngredients(recipeCreateReq.getIngredients());
+            Board.get().setThumbnailImageUrl(thumbnailImg);
+
+            boardService.save(Board.get());  // 단계들 빼고 글, 사진 DB에 올리고 게시글 번호 반환
+
+            // 기존 레시피 단계들 다 삭제
+            List<boardSteps> steps = boardStepsService.findAllByBoardIdOrderByIdAsc(Board.get());
+            for(int i = 0; i < steps.size(); i++) {
+                awsS3Service.deleteImage(steps.get(i).getImageUrl());  //  S3에 올렸던 단계별 이미지들 다 삭제
+            }
+            boardStepsService.deleteAllByBoardId(Board.get());  // 단계들 다 삭제
+
+            List<String> ImgNameList = awsS3Service.uploadImages(file);  // 레시피 단계 사진들 S3에 게시
+
+            boardStepsService.save(recipeCreateReq.getRecipe_steps(), Board.get(), ImgNameList, recipeCreateReq.getStepNums());  // 레시피 단계 글이랑 사진 DB에 저장
 
             return ResponseEntity.status(200).body(BaseResponseBody.of(200, "success"));
         } catch (Exception e) {
@@ -79,7 +133,13 @@ public class BoardController {
         try {
             UserDto userDto = userService.getUserByToken(request.getHeader("Authorization").substring(7));
             Optional<board> Board = boardService.findByIdAndUserEmail(id, userDto.toEntity());
-            if(Board.isPresent()) {
+            if(!Board.isPresent())  // id에 맞는 게시글이 없으면 null리턴
+                return ResponseEntity.status(200).body(BaseResponseBody.of(401, "There's no such BoardId"));
+            else {
+                if(Board.get().getUserEmail().getEmail() != userDto.getEmail()) {  // 본인이 작성한 레시피가 아니라면
+                    return ResponseEntity.status(200).body(BaseResponseBody.of(401, "Unauthorized"));
+                }
+
                 boardService.deleteById(id);
             }
             return ResponseEntity.status(200).body(BaseResponseBody.of(200, "success"));
@@ -97,8 +157,10 @@ public class BoardController {
             @ApiResponse(code = 404, message = "사용자 없음"),
             @ApiResponse(code = 500, message = "서버 오류")
     })
-    public ResponseEntity<? extends Object> selectAllRecipe() throws IOException {
+    public ResponseEntity<? extends Object> selectAllRecipe(HttpServletRequest request) throws IOException {
         try {
+            UserDto userDto = userService.getUserByToken(request.getHeader("Authorization").substring(7));
+
             List<BoardDto> Boards = boardService.findAll();
             List<SelectAllRecipeRes> Recipes = new ArrayList<>();
             for(int i = 0; i < Boards.size(); i++) {
@@ -140,7 +202,7 @@ public class BoardController {
 
                 // 사용자가 좋아요 했는지 여부를 판단 위함 + 토큰 여부
                 int tokenChk = 0;   // 토큰 없음
-                if(request.getHeader("Bearer") == null) {
+                if(request.getHeader("Authorization") == null) {
                     selectOneRecipeRes.setLikeOrNot(0);  // 좋아요 체크 안한상태로 보냄
                 } else {
                     tokenChk = 1;
@@ -234,6 +296,9 @@ public class BoardController {
             if(!Board.isPresent()) {
                 return ResponseEntity.status(400).body(BaseResponseBody.of(400, "There's no such BoardId"));
             }
+            if(userDto.getEmail() != commentService.findById(id).get().getUserEmail().getEmail()) {  // 본인이 아니라면
+                return ResponseEntity.status(400).body(BaseResponseBody.of(401, "UnAuthorized"));
+            }
             Optional<likeTable> like = likeService.findByBoardIdAndUserEmail(Board.get(), userDto.toEntity());
             likeService.deleteById(like.get().getId());
             return ResponseEntity.status(200).body(BaseResponseBody.of(200, "success"));
@@ -260,13 +325,42 @@ public class BoardController {
                 return ResponseEntity.status(400).body(BaseResponseBody.of(401, "There's no such BoardId"));
             }
             CommentDto commentDto = new CommentDto(userDto.toEntity(), Board.get(), commentReq.getContent(), LocalDateTime.now());
-            commentService.save(commentDto);
+            commentService.save(commentDto.toEntity());
             return ResponseEntity.status(200).body(BaseResponseBody.of(200, "success"));
         } catch (Exception e) {
             e.printStackTrace();
             return ResponseEntity.status(401).body(BaseResponseBody.of(401, "Bad Request"));
         }
     }
+
+    @PatchMapping("/comments/{commentId}")
+    @ApiOperation(value = "게시판 댓글 수정", notes = "<strong>게시판 댓글 수정</strong>")
+    @ApiResponses({
+            @ApiResponse(code = 200, message = "성공"),
+            @ApiResponse(code = 204, message = "No Content"),
+            @ApiResponse(code = 401, message = "인증 실패"),
+            @ApiResponse(code = 404, message = "사용자 없음"),
+            @ApiResponse(code = 500, message = "서버 오류")
+    })
+    public ResponseEntity<? extends Object> patchCommentBoard(@PathVariable("commentId") Long id, @RequestBody @ApiParam(value="댓글 내용", required = true) CommentReq commentReq, HttpServletRequest request) throws IOException {
+        try {
+            UserDto userDto = userService.getUserByToken(request.getHeader("Authorization").substring(7));
+            Optional<comment> comment = commentService.findById(id);
+            if(!comment.isPresent()) {
+                return ResponseEntity.status(400).body(BaseResponseBody.of(401, "There's no such commentId"));
+            }
+            if(userDto.getEmail() != commentService.findById(id).get().getUserEmail().getEmail()) {  // 본인이 아니라면
+                return ResponseEntity.status(400).body(BaseResponseBody.of(401, "UnAuthorized"));
+            }
+            comment.get().setContent(commentReq.getContent());
+            commentService.save(comment.get());
+            return ResponseEntity.status(200).body(BaseResponseBody.of(200, "success"));
+        } catch (Exception e) {
+            e.printStackTrace();
+            return ResponseEntity.status(401).body(BaseResponseBody.of(401, "Bad Request"));
+        }
+    }
+
 
     @DeleteMapping("/comments/{commentId}")
     @ApiOperation(value = "게시판 댓글 삭제", notes = "<strong>게시판 댓글 삭제</strong>")
@@ -279,8 +373,13 @@ public class BoardController {
     })
     public ResponseEntity<? extends Object> deleteCommentBoard(@PathVariable("commentId") Long id, HttpServletRequest request) throws IOException {
         try {
+            UserDto userDto = userService.getUserByToken(request.getHeader("Authorization").substring(7));
+
             if(!commentService.findById(id).isPresent()) {
                 return ResponseEntity.status(400).body(BaseResponseBody.of(400, "There's no such BoardId"));
+            }
+            if(userDto.getEmail() != commentService.findById(id).get().getUserEmail().getEmail()) {
+                return ResponseEntity.status(400).body(BaseResponseBody.of(401, "UnAuthorized"));
             }
             commentService.deleteById(id);
             return ResponseEntity.status(200).body(BaseResponseBody.of(200, "success"));
@@ -299,8 +398,9 @@ public class BoardController {
             @ApiResponse(code = 404, message = "사용자 없음"),
             @ApiResponse(code = 500, message = "서버 오류")
     })
-    public ResponseEntity<? extends Object> getTopRecipes() throws IOException {
+    public ResponseEntity<? extends Object> getTopRecipes(HttpServletRequest request) throws IOException {
         try {
+            UserDto userDto = userService.getUserByToken(request.getHeader("Authorization").substring(7));
             List<SelectAllRecipeRes> boards = boardService.findTop3ByOrderByLikesDesc();
             return ResponseEntity.status(200).body(boards);
         } catch (Exception e) {
@@ -317,8 +417,9 @@ public class BoardController {
             @ApiResponse(code = 404, message = "사용자 없음"),
             @ApiResponse(code = 500, message = "서버 오류")
     })
-    public ResponseEntity<? extends Object> getBoardCuisineOrderByDate(@PathVariable("cuisineName") String title) throws IOException {
+    public ResponseEntity<? extends Object> getBoardCuisineOrderByDate(@PathVariable("cuisineName") String title, HttpServletRequest request) throws IOException {
         try {
+            UserDto userDto = userService.getUserByToken(request.getHeader("Authorization").substring(7));
             List<BoardDto> Boards = boardService.findByTitleContainingOrderByCreateDate(title);
             List<SelectAllRecipeRes> Recipes = new ArrayList<>();
             for(int i = 0; i < Boards.size(); i++) {
@@ -342,8 +443,9 @@ public class BoardController {
             @ApiResponse(code = 404, message = "사용자 없음"),
             @ApiResponse(code = 500, message = "서버 오류")
     })
-    public ResponseEntity<? extends Object> getBoardCuisineOrderByLikes(@PathVariable("cuisineName") String title) throws IOException {
+    public ResponseEntity<? extends Object> getBoardCuisineOrderByLikes(@PathVariable("cuisineName") String title, HttpServletRequest request) throws IOException {
         try {
+            UserDto userDto = userService.getUserByToken(request.getHeader("Authorization").substring(7));
             List<BoardDto> Boards = boardService.findByTitleContainingOrderByLikesDesc(title);
             List<SelectAllRecipeRes> Recipes = new ArrayList<>();
             for(int i = 0; i < Boards.size(); i++) {
